+++ rust/stackablectl/src/args/mod.rs
@@ -1,9 +1,11 @@
 mod cluster;
 mod file;
 mod namespace;
+mod operator_configs;
 mod repo;
 
 pub use cluster::*;
 pub use file::*;
 pub use namespace::*;
+pub use operator_configs::*;
 pub use repo::*;